Moved the WiFi driver instruction to the very end
I assume, that I may not be the only one expecting the most recent official build on elementaryos.org, instead of some sourceforge repository. Thus I personally think this could still be of value for some people, until the link in the journal article gets updated.
But maybe that's just me...

In the end it's completely up to you, whether or not to include it.
I just wanted to make one last proposal ;-)